feat: @stt/deck.gl v2 layers (GPU time filter, consolidation, trips)
Update the deck.gl layer family for the v2 BinaryFeatures shape:

- SpatioTemporalLayer: owns STTArchive + SpatiotemporalTileset, wires
  finalize() through the deck.gl layer lifecycle so worker pools and the
  archive cache are released on layer unmount.

- AnimatedPointLayer / AnimatedPathLayer / AnimatedPolygonLayer /
  AnimatedTripsLayer: consume consolidated buffers, accept dynamic
  getTime() so the layer stays cached while only uniforms update per
  frame.

- TimeFilterExtension: explicit f32 precision contract — attribute times
  are relative to a per-layer timeOffset; the extension subtracts the
  same offset from currentTime before setting the uniform. Window and
  trail modes share one shader path; MAX_RELATIVE_TIME_MS guards the
  contract with a one-shot warning.

- consolidate.ts: new file that merges several tiles' BinaryFeatures
  into one set of typed arrays for a single deck.gl draw call. Rebases
  per-tile time offsets to a single layer-wide offset so f32 stays
  exact even when tiles span the boundary.

- CategoryColorExtension: texture-based palette lookup so a categorical
  property does not need CPU-side color expansion.

- TimeController: requestAnimationFrame playback clock, loop modes,
  speed control, tick subscribers.

- Tests cover consolidation correctness, f32 precision, attribute
  wiring, time-controller behaviour, and a 200k-feature perf budget.

+++ packages/deck.gl/src/heatmap-time-layer.ts
@@ -18,12 +18,17 @@ import type { BinaryFeatures, Tile } from '@stt/core';
 
 /**
  * Cached point data from a single tile layer (all points, not time-filtered)
+ *
+ * Times are stored as plain JS-number arrays of ABSOLUTE epoch-ms. They are
+ * NOT Float32Array: absolute epoch-ms (~1.7e12) overflows float32's 24-bit
+ * mantissa (quantizing to ~131s buckets). Filtering here happens on the CPU
+ * in JS doubles, so plain number[] keeps full precision.
  */
 interface CachedTilePoints {
   positions: Float64Array;  // Interleaved [lon, lat, lon, lat, ...]
   weights: Float32Array;
-  startTimes: Float32Array; // Absolute times
-  endTimes: Float32Array;   // Absolute times
+  startTimes: Float64Array; // Absolute times (epoch-ms) - Float64 for precision
+  endTimes: Float64Array;   // Absolute times (epoch-ms) - Float64 for precision
   count: number;
 }
 
@@ -242,8 +247,9 @@ export class HeatmapTimeLayer extends SpatioTemporalLayer<HeatmapTimeLayerProps>
     
     const positions = new Float64Array(binary.featureCount * 2);
     const weights = new Float32Array(binary.featureCount);
-    const startTimes = new Float32Array(binary.featureCount);
-    const endTimes = new Float32Array(binary.featureCount);
+    // Float64 so absolute epoch-ms keeps full precision (see CachedTilePoints).
+    const startTimes = new Float64Array(binary.featureCount);
+    const endTimes = new Float64Array(binary.featureCount);
     
     for (let i = 0; i < binary.featureCount; i++) {
       const posIdx = i * dims;
